Pause timer and testing memory capabilities
In this update I have create a pause timer to pause the game, meaning all entities including the player are paused. The system will not read any inputs given by the user except the pause key within the paused state to return to the un-pause state.

Changing the number of enemy entities to test the amount of spare memory can be allocated withing the system as well how its correlates to its spawn rate.